added search controller
